+++ src/LoginPage.jsx
@@ -1,50 +1,10 @@
-import { useDispatch } from 'react-redux';
-import { requestLogin, updateLoginFields } from './store';
+import LoginFormContainer from './LoginFormContainer';
 
-export default function LoginPage({ email, password }) {
-  const dispatch = useDispatch();
-
-  function handleChange(event) {
-    const { name, value } = event.target;
-
-    dispatch(updateLoginFields({ name, value }));
-  }
-
-  function handleClick() {
-    dispatch(requestLogin());
-  }
-
+export default function LoginPage() {
   return (
     <div>
       <h1>Login</h1>
-      <div>
-        <label htmlFor="login-email">email</label>
-        <input
-          type="email"
-          name="email"
-          placeholder="email"
-          id="login-email"
-          value={email}
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <label htmlFor="login-password">password</label>
-        <input
-          type="password"
-          name="password"
-          placeholder="password"
-          id="login-password"
-          value={password}
-          onChange={handleChange}
-        />
-      </div>
-      <button
-        type="button"
-        onClick={handleClick}
-      >
-        Log In
-      </button>
+      <LoginFormContainer />
     </div>
   );
 }
