+++ src/LogoutForm.jsx
@@ -1,0 +1,7 @@
+export default function LogoutForm() {
+  return (
+    <div>
+      <button type="button">Logout</button>
+    </div>
+  );
+}
